Refine StatusDropdown
