+++ src/components/StatusDropdown.js
@@ -1,5 +1,8 @@
 import React, { Component, PropTypes } from 'react';
 
+import { powerOnLinode, powerOffLinode, rebootLinode } from '~/api/linodes';
+import { LinodeStates, LinodeStatesReadable } from '~/constants';
+
 export default class StatusDropdown extends Component {
   constructor() {
     super();
@@ -19,7 +22,37 @@ export default class StatusDropdown extends Component {
   }
 
   render() {
-    const dropdownMenu = this.props.elements.map(({ name, action, _key }) =>
+    const { linode, dispatch } = this.props;
+    const dropdownElements = [
+      {
+        name: <span><i className="fa fa-refresh"></i> Reboot</span>,
+        _key: 'reboot',
+        _action: rebootLinode,
+        _condition: () => linode.status !== 'offline',
+      },
+      {
+        name: <span><i className="fa fa-power-off"></i> Power off</span>,
+        _key: 'power-off',
+        _action: powerOffLinode,
+        _condition: () => linode.status === 'running',
+      },
+      {
+        name: <span><i className="fa fa-power-off"></i> Power on</span>,
+        _key: 'power-on',
+        _action: powerOnLinode,
+        _condition: () => linode.status === 'offline',
+      },
+    ]
+    .filter(element => element._condition())
+    .map(element => ({
+      ...element,
+      action: () => {
+       this.close();
+       dispatch(element._action(linode.id, this.state.config || null));
+      }
+    }));
+
+    const dropdownMenu = dropdownElements.map(({ name, action, _key }) =>
       <button
         type="button"
         key={_key}
@@ -37,7 +70,9 @@ export default class StatusDropdown extends Component {
         onBlur={this.close}
       >
         <div className="status-dropdown pull-left">
-          {this.props.children}
+          <span className={`linode-status ${linode.status}`}>
+            {LinodeStatesReadable[linode.status]}
+          </span>
         </div>
         <button
           type="button"
@@ -46,6 +81,7 @@ export default class StatusDropdown extends Component {
           aria-haspopup="true"
           aria-expanded={this.state.open}
           onClick={this.open}
+          disabled={LinodeStates.pending.indexOf(linode.status) !== -1}
         >
           <span className="sr-only">Toggle dropdown</span>
         </button>
@@ -56,9 +92,7 @@ export default class StatusDropdown extends Component {
 }
 
 StatusDropdown.propTypes = {
-  elements: PropTypes.arrayOf(PropTypes.shape({
-    name: PropTypes.node.isRequired,
-    action: PropTypes.func,
-  })).isRequired,
+  linode: PropTypes.object,
+  dispatch: PropTypes.func,
   leftOriented: PropTypes.bool,
 };
